+++ client/client-app/src/modules/AddTags.js
@@ -66,7 +66,9 @@ export default function AddTags({tagContent}) {
             
             <div className="fill">
               <div className="input-container text-box ql-container">
+              <h4>Press Enter To Submit</h4>
               <p>Make sure to make them accurate and clear as possible</p>
+          
            
            
             <input
